Replace use of AppCache PROJECT section with strongly-typed structures

+++ src/main.cpp
@@ -2798,9 +2798,11 @@ bool CBlock::DisconnectBlock(CTxDB& txdb, CBlockIndex* pindex)
                 
                 try
                 {
-                DeleteCache(StringToSection(sMType), sMKey);
-                if(fDebug)
-                    LogPrintf("DisconnectBlock: Delete contract %s %s", sMType, sMKey);
+                    if (!NN::DeleteContract(sMType, sMKey)) {
+                        DeleteCache(StringToSection(sMType), sMKey);
+                    }
+                    if(fDebug)
+                        LogPrintf("DisconnectBlock: Delete contract %s %s", sMType, sMKey);
                 }
                 catch(const std::runtime_error& e)
                 {
@@ -8094,10 +8096,13 @@ bool MemorizeMessage(const CTransaction &tx, double dAmount, std::string sRecipi
 
                     try
                     {
-                        WriteCache(StringToSection(sMessageType), sMessageKey,sMessageValue,nTime);
-                        if(fDebug10 && sMessageType=="beacon" )
-                                    LogPrintf("BEACON add %s %s %s", sMessageKey, DecodeBase64(sMessageValue), TimestampToHRDate(nTime));
-                                }
+                        if (!NN::AddContract(sMessageType, sMessageKey, sMessageValue, nTime)) {
+                            WriteCache(StringToSection(sMessageType), sMessageKey,sMessageValue,nTime);
+
+                            if(fDebug10 && sMessageType=="beacon" )
+                                LogPrintf("BEACON add %s %s %s", sMessageKey, DecodeBase64(sMessageValue), TimestampToHRDate(nTime));
+                        }
+                    }
                     catch(const std::runtime_error& e)
                     {
                         error("Attempting to add to unknown cache: %s", sMessageType);
@@ -8118,9 +8123,11 @@ bool MemorizeMessage(const CTransaction &tx, double dAmount, std::string sRecipi
                     
                     try
                     {
-                        DeleteCache(StringToSection(sMessageType), sMessageKey);
-                                fMessageLoaded = true;
+                        if (!NN::DeleteContract(sMessageType, sMessageKey)) {
+                            DeleteCache(StringToSection(sMessageType), sMessageKey);
                         }
+                        fMessageLoaded = true;
+                    }
                     catch(const std::runtime_error& e)
                     {
                         error("Attempting to add to unknown cache: %s", sMessageType);
